+++ src/Item/index.ts
@@ -1,0 +1,2 @@
+export { SpinoramaItem } from './Item'
+export * from './Item.types'