+++ src/Slideshow/Slideshow.tsx
@@ -1,34 +1,30 @@
 import * as React from 'react'
-import { Grid } from '@mui/material'
+import { Box } from '@mui/material'
 
 // Types
 import type { SpinoramaSlideshowProps } from './Slideshow.types'
+
+// Components
+import { SpinoramaItem } from '../Item'
 
 const SpinoramaSlideshow: React.FC<SpinoramaSlideshowProps> = (props: SpinoramaSlideshowProps) => {
 	// Props
 	const { className, children } = props
 
-	// Check if children is defined
-	if (!children) return null
-
 	return (
-		<Grid
+		<Box
 			className={`spinorama-slidehow${className ? ` ${className}` : ''}`}
+			display="flex"
 			width={1}
 			overflow="hidden"
 			sx={{
 				tabSize: React.Children.count(children) ?? 0
 			}}>
 			{React.Children.map(children, (child, index) => {
-				if (!React.isValidElement(child)) return null
-
-				return (
-					<Grid key={index} className="spinorama-slidehow-item" flex="0 0 100%">
-						{React.cloneElement(child)}
-					</Grid>
-				)
+				if (React.isValidElement(child)) return <SpinoramaItem key={index}>{React.cloneElement(child)}</SpinoramaItem>
+				else return null
 			})}
-		</Grid>
+		</Box>
 	)
 }