+++ src/index.d.ts
@@ -1,2 +1,3 @@
 export * from './Spinorama'
 export * from './Slideshow'
+export * from './Item'